+++ src/Frontend/Dfe.Complete/Pages/Projects/TaskList/Tasks/BaseProjectTaskModel.cs
@@ -54,6 +54,7 @@ public class BaseProjectTaskModel(ISender sender, IAuthorizationService authoriz
        NoteTaskIdentifier.Subleases,
        NoteTaskIdentifier.TenancyAtWill,
        NoteTaskIdentifier.ThirdPartyLeases,
+    NoteTaskIdentifier.ConfirmSchoolBankDetails,
        NoteTaskIdentifier.ConfirmSchoolHasCompletedAllActions,       
        NoteTaskIdentifier.ShareInformationAboutOpening,
        NoteTaskIdentifier.ConfirmAcademyOpenedDate,