+++ src/Frontend/Dfe.Complete/Pages/Projects/TaskList/Tasks/ConfirmSchoolBankDetailsTask/ConfirmSchoolBankDetailsTask.cshtml.cs
@@ -1,0 +1,45 @@
+using Dfe.Complete.Application.Projects.Commands.TaskData;
+using Dfe.Complete.Constants;
+using Dfe.Complete.Domain.Enums;
+using Dfe.Complete.Domain.ValueObjects;
+using Dfe.Complete.Services;
+using MediatR;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+
+namespace Dfe.Complete.Pages.Projects.TaskList.Tasks.ConfirmSchoolBankDetailsTask
+{
+    public class ConfirmSchoolBankDetailsTaskModel(ISender sender, IAuthorizationService authorizationService, ILogger<ConfirmSchoolBankDetailsTaskModel> logger, IProjectPermissionService projectPermissionService)
+    : BaseProjectTaskModel(sender, authorizationService, logger, NoteTaskIdentifier.ConfirmSchoolBankDetails, projectPermissionService)
+    {
+        [BindProperty]
+        public Guid? TasksDataId { get; set; }
+
+        [BindProperty(Name = "sent")]
+        public bool? Sent { get; set; }
+
+        [BindProperty(Name = "submitted")]
+        public bool? Submitted { get; set; }
+
+        public override async Task<IActionResult> OnGetAsync()
+        {
+            await base.OnGetAsync();
+
+            if (InvalidTaskRequestByProjectType())
+                return Redirect(RouteConstants.ErrorPage);
+
+            TasksDataId = Project.TasksDataId?.Value;
+            Sent = ConversionTaskData.ConfirmSchoolBankDetailsSent;
+            Submitted = ConversionTaskData.ConfirmSchoolBankDetailsSubmitted;
+
+            return Page();
+        }
+
+        public async Task<IActionResult> OnPost()
+        {
+            await Sender.Send(new UpdateConfirmSchoolBankDetailsTaskCommand(new TaskDataId(TasksDataId.GetValueOrDefault())!, Sent, Submitted));
+            SetTaskSuccessNotification();
+            return Redirect(string.Format(RouteConstants.ProjectTaskList, ProjectId));
+        }
+    }
+}