+++ src/Tests/Dfe.Complete.CypressTests/cypress/api/taskApiConversions.ts
@@ -44,6 +44,12 @@ interface UpdateConfirmSchoolHasCompletedAllActionsTaskRequest {
     saved?: boolean;
 }
 
+interface UpdateConfirmSchoolBankDetailsTaskRequest {
+    taskDataId: TaskDataId;
+    sent?: boolean;
+    submitted?: boolean;
+}
+
 interface UpdateDirectionToTransferTaskRequest {
     taskDataId: TaskDataId;
     notApplicable?: boolean;
@@ -179,6 +185,10 @@ class TaskApiConversions extends TaskApi {
         requestBody: UpdateConfirmSchoolHasCompletedAllActionsTaskRequest,
     ) {
         return this.taskDataBaseRequest<void>("UpdateConfirmSchoolHasCompletedAllActions", requestBody);
+    }
+
+    public updateConfirmSchoolBankDetailsTask(requestBody: UpdateConfirmSchoolBankDetailsTaskRequest) {
+        return this.taskDataBaseRequest<void>("ConfirmSchoolBankDetails", requestBody);
     }
 
     public updateDirectionToTransferTask(requestBody: UpdateDirectionToTransferTaskRequest) {