+++ src/Tests/Dfe.Complete.CypressTests/cypress/pages/projects/conversionTaskListPage.ts
@@ -43,6 +43,7 @@ class ConversionTaskListPage {
             "Commercial transfer agreement",
         ],
         readyForOpening: [
+            "Confirm the new bank account details for the school",
             "Confirm the school has completed all actions",
             "Confirm all conditions have been met",
             "Share the information about opening",